turn on some additional warning and fix them ;-)
git-svn-id: svn://svn.freaque.net/enigma2@851 a295cab0-163a-0410-857a-8207b47ee9d6

+++ Classes/SmallSockets/BufferedSocket.h
@@ -68,7 +68,7 @@
 
 // Reading and writing data
 
-- (NSMutableData*)readData: (int)n;
+- (NSMutableData*)readData: (NSUInteger)n;
 - (NSMutableData*)readDataUpToData:(NSData*) d;
 - (NSMutableData*)readDataUpToString:(NSString*) s;
